@@ -19,7 +19,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.18.1 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -57,6 +57,41 @@ dependencies:
   full_screen_image: ^1.0.2
   image_viewer: ^1.5.0
   easy_image_viewer: ^1.2.0
+  # flutter_svg: ^2.0.4
+  emoji_picker_flutter: ^1.5.2
+  # emoji_picker: ^0.1.0
+  camera: ^0.10.3+2
+  path_provider: ^2.0.14
+  path:
+  video_player: ^2.6.1
+  # socket_io_client: ^2.0.1
+  socket_io_client: ^1.0.1
+
+
+  avoid_keyboard: ^0.3.0
+  # http: ^0.13.5
+  flutter_redux: ^0.10.0
+  axios: ^1.0.1+1
+  localstorage: ^4.0.1+2
+  # intl: ^0.18.0
+  uuid: ^3.0.7
+  convert: ^3.1.1
+  redux_thunk: ^0.4.0
+  flutter_lints: ^2.0.0
+  google_nav_bar: ^5.0.6
+  line_icons: ^2.0.1
+  flutter_launcher_icons: "^0.9.2"
+  
+
+  animated_text_kit: ^4.2.2
+  # cupertino_icons: ^1.0.2
+  # font_awesome_flutter: ^10.4.0
+  
+  flutter_animated_button: ^2.0.3
+  google_fonts: ^4.0.3
+  flutter_switch: ^0.3.2
+ 
+  
 
 
 dev_dependencies:
@@ -68,7 +103,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
+  # flutter_lints: ^1.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -85,6 +120,15 @@ flutter:
     - assets/images/
     - assets/applepay.json
     - assets/gpay.json
+    - assets/images/ICU.jpg
+    - assets/images/Ventilator.jpg
+    - assets/images/General.jpg
+    - assets/images/VIP.jpg
+    - assets/images/Grande.jpeg
+    - assets/images/Norvic.jpg
+    - assets/images/Dhulikhel.jpg
+    - assets/images/greencity.jpg
+    - assets/images/cmc.jpg
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
@@ -109,6 +153,13 @@ flutter:
   #       - asset: fonts/TrajanPro.ttf
   #       - asset: fonts/TrajanPro_Bold.ttf
   #         weight: 700
-  #
+    #  fonts:
+    # - family: OpenSans
+    #   fonts:
+    #     - asset: fonts/OpenSans-Bold.ttf
+    #     - asset: fonts/OpenSans-Regular.ttf
+    #     - asset: fonts/OpenSans-SemiBold.ttf
+    #     - asset: fonts/OpenSans-Italic.ttf
+    #       style: italic
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
