@@ -1,11 +1,6 @@
 // import 'dart:convert';
 
-// import 'package:amazon_clone_tutorial/constants/error_handling.dart';
-// import 'package:amazon_clone_tutorial/constants/global_variables.dart';
-// import 'package:amazon_clone_tutorial/constants/utils.dart';
-// import 'package:amazon_clone_tutorial/features/home/widgets/newproduct.dart';
-// import 'package:amazon_clone_tutorial/models/product.dart';
-// import 'package:amazon_clone_tutorial/providers/user_provider.dart';
+
 // import 'package:flutter/material.dart';
 // import 'package:provider/provider.dart';
 // import 'package:http/http.dart' as http;
@@ -278,6 +273,38 @@ class HomeServices {
       showSnackBar(context, e.toString());
     }
     return prescriptionList;
+  }
+// fetch all user
+ 
+  Future<List<User>> fetchAllUsers(BuildContext context) async {
+    final userProvider = Provider.of<UserProvider>(context, listen: false);
+    List<User> userList = [];
+    try {
+      http.Response res =
+          await http.get(Uri.parse('$uri/admin/get-users'), headers: {
+        'Content-Type': 'application/json; charset=UTF-8',
+        'x-auth-token': userProvider.user.token,
+      });
+
+      httpErrorHandle(
+        response: res,
+        context: context,
+        onSuccess: () {
+          for (int i = 0; i < jsonDecode(res.body).length; i++) {
+            userList.add(
+              User.fromJson(
+                jsonEncode(
+                  jsonDecode(res.body)[i],
+                ),
+              ),
+            );
+          }
+        },
+      );
+    } catch (e) {
+      showSnackBar(context, e.toString());
+    }
+    return userList;
   }
 
   Future<List<Pres>> fetchMyPrescription({
