@@ -1,3 +1,4 @@
+import 'package:luveen/models/Doctor.dart';
 import 'package:luveen/models/user.dart';
 import 'package:flutter/material.dart';
 
@@ -9,17 +10,22 @@ class UserProvider extends ChangeNotifier {
     password: '',
     address: '',
     contactno: '',
-    type: '',
+    role: '',
     token: '',
     cart: [],
     prescription: [],
     profiles:[],
+    doctors:[],
+    // doctor: Doctor(name: '', email: '', password: '', contactno: '', address: '', qualification: '', specialization: '', experience: '', nmc_no: 0) ,
   );
 
   User get user => _user;
 
+  get doctor => null;
+
   void setUser(String user) {
     _user = User.fromJson(user);
+    // _user = User.fromJson(user as Map<String, dynamic>);
     notifyListeners();
   }
 
