@@ -16,7 +16,7 @@ const admin = async (req, res, next) => {
         .status(401)
         .json({ msg: "Token verification failed, authorization denied." });
     const user = await User.findById(verified.id);
-    if (user.type == "user" || user.type == "seller") {
+    if (user.role == "user" || user.role == "seller") {
       return res.status(401).json({ msg: "You are not an admin!" });
     }
     req.user = verified.id;
